# OpenAPI.yaml
openapi: 3.0.0
info:
  title: MY SWEET SKRIPSI
  contact: {}
  version: "1.0.0"
servers:
  - url: http://localhost:8000/v1/api
    variables: {}
paths:
...
  /kriteria/byid/{id}:
    get:
      tags:
        - For Data Kriteria and Sub-kriteria
      summary: GET DATA KRITERIA BY ID
      description: ""
      operationId: GETDATAKRITERIABYID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful Get data kriteria by id
          headers: {}
          content: {}
        "400":
          description: Error or not success get data kriteria by id
          headers: {}
          content: {}
        "500":
          description: internal server error
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearerAuth: []
...
  /action/calculatedROC:
    post:
      tags:
        - For Calculated
      summary: CALCULATED ROC
      description: ""
      operationId: CALCULATEDROC
      parameters: []
      responses:
        "200":
          description: Successful calculated ROC
          headers: {}
          content: {}
        "400":
          description: Error or not success calculated ROC
          headers: {}
          content: {}
        "500":
          description: internal server error
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security: []
...